refactor: 댓글 목록 조회 코드 서비스 파일 위치 수정
